+++ caffeApp/IProduct.cs
@@ -9,6 +9,6 @@ namespace VubCaffe
     public interface IProduct
     {
         string Name { get; }
-        decimal Price { get; }
+        double Price { get; }
     }
 }
